# openapi.yaml
openapi: 3.0.3
info:
  title: M321 Order Service API
  version: 5.0.0
  description: REST entry point for orders. Processing is delegated asynchronously via RabbitMQ.
servers:
  - url: http://localhost:8081
paths:
  /api/v1/orders:
    post:
      operationId: createOrderV1
      summary: Accept an order using the stable v1 request
      security:
        - bearerJwt: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderV1Request'
            example:
              product: Keyboard
              quantity: 2
      responses:
        '202':
          description: Order accepted and OrderCreated v1 published
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAcceptedResponse'
              example:
                orderId: 6fe4b9bb-d89d-43c2-af6a-cd6fa4b453ae
                correlationId: demo-correlation-123
                status: ACCEPTED
                acceptedAt: '2026-09-18T08:00:00Z'
        '400':
          description: Product is blank or quantity is below one
        '401':
          description: Missing or invalid OAuth2 access token
        '403':
          description: Token has neither USER nor ADMIN role
  /api/v2/orders:
    post:
      operationId: createOrderV2
      summary: Accept an order using the new, incompatible v2 request shape
      description: Runs in parallel with v1 during the expand-and-contract migration.
      security:
        - bearerJwt: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderV2Request'
            example:
              productCode: KEYBOARD-01
              amount: 2
      responses:
        '202':
          description: Order accepted and OrderCreated v1 published
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAcceptedResponse'
        '400':
          description: Product code is blank or amount is below one
        '401':
          description: Missing or invalid OAuth2 access token
        '403':
          description: Token has neither USER nor ADMIN role
  /api/admin:
    get:
      operationId: getAdminInfo
      summary: Verify ADMIN access
      security:
        - bearerJwt: []
      responses:
        '200':
          description: ADMIN access granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInfo'
        '401':
          description: Missing or invalid OAuth2 access token
        '403':
          description: Token does not contain ADMIN
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 access token issued by the m321 Keycloak realm
    keycloakOidc:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8080/realms/m321/.well-known/openid-configuration
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Existing workflow identifier. A UUID is generated when omitted.
      schema:
        type: string
      example: demo-correlation-123
  headers:
    CorrelationId:
      description: Effective correlation ID, supplied by the client or generated by the service
      schema:
        type: string
  schemas:
    CreateOrderV1Request:
      type: object
      additionalProperties: false
      required: [product, quantity]
      properties:
        product:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
    CreateOrderV2Request:
      type: object
      additionalProperties: false
      required: [productCode, amount]
      properties:
        productCode:
          type: string
          minLength: 1
        amount:
          type: integer
          minimum: 1
    OrderAcceptedResponse:
      type: object
      required: [orderId, correlationId, status, acceptedAt]
      properties:
        orderId:
          type: string
          format: uuid
        correlationId:
          type: string
        status:
          type: string
          enum: [ACCEPTED]
        acceptedAt:
          type: string
          format: date-time
    AdminInfo:
      type: object
      required: [subject, username, message]
      properties:
        subject:
          type: string
        username:
          type: string
        message:
          type: string